refactor: Move EntrySnapshot and BatchUpdate definitions to the stats module
